api/requirements.txt
supabase==2.3.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic==2.5.0
